Address the second review round on the stripe-mock starter
Moves the stand-in mock binaries out of the shared helpers and in beside the
suite that uses them, which puts the shared file back under the size limit.

Makes two tests stop depending on which process the machine happens to run
next. Looking for an open port now goes through a stand-in that counts the
looks and can refuse any of them, so the test named for looking twice refuses
the first look outright rather than racing a sleep, and the mock that dies
while being checked waits to be looked at before it starts dying. Both would
have passed either way before, quietly testing something else.

Counting the looks also made the polling pause testable, so the mutant recorded
as impossible to catch is now caught by a test instead: a mock nothing ever
answers, given three hundred milliseconds, is looked for a few dozen times
rather than many thousands. The file is now fully covered with nothing
suppressed.
